Move EntityPath and NameLookup to core
