+++ crates/bevy_core/Cargo.toml
@@ -17,6 +17,7 @@ bevy_math = { path = "../bevy_math", version = "0.9.0-dev" }
 bevy_reflect = { path = "../bevy_reflect", version = "0.9.0-dev", features = ["bevy"] }
 bevy_tasks = { path = "../bevy_tasks", version = "0.9.0-dev" }
 bevy_utils = { path = "../bevy_utils", version = "0.9.0-dev" }
+bevy_hierarchy = { path = "../bevy_hierarchy", version = "0.9.0-dev" }
 
 # other
 bytemuck = "1.5"
